requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1

# Optional: For legacy .doc file support (MS Word 97-2003)
# textract==1.6.5
# Note: On Windows, textract may require additional setup (antiword, etc.)
